Separate styles from header index
